Remove PyQt6 dependencies from services/report_generator.py (#629)
Split ReportGenerator into two layers:
- services/report_generator.py: pure Python ReportConfig, ReportData,
  and ReportDataBuilder (zero Qt imports)
- GUI/report_renderer.py: PDFReportRenderer with all Qt rendering logic

This eliminates the backwards dependency (services/ -> GUI/) and allows
the services layer to be used and tested without PyQt6.
